# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KlikFilm Scraper API - Film & Anime
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type MovieType = (typeof MovieType)[keyof typeof MovieType];

export const MovieType = {
  film: "film",
  anime: "anime",
  all: "all",
} as const;

export interface Movie {
  title: string;
  slug: string;
  poster: string;
  year?: string;
  rating?: string;
  genre?: string[];
  quality?: string;
  type?: MovieType;
  url: string;
}

export type MovieDetailType =
  (typeof MovieDetailType)[keyof typeof MovieDetailType];

export const MovieDetailType = {
  film: "film",
  anime: "anime",
  all: "all",
} as const;

export interface MovieDetail {
  title: string;
  slug: string;
  poster: string;
  year?: string;
  rating?: string;
  genre?: string[];
  quality?: string;
  type?: MovieDetailType;
  url: string;
  synopsis?: string;
  duration?: string;
  country?: string;
  director?: string;
  cast?: string[];
  /** Direct video URLs (MP4/M3U8), ad-free */
  streamUrls?: string[];
}

export interface MoviesResponse {
  movies: Movie[];
  totalPages: number;
  currentPage: number;
}

export interface GenresResponse {
  genres: string[];
}

export type ScraperStatusLastRunResult = {
  films?: number;
  anime?: number;
} | null;

export interface ScraperStatus {
  lastRunAt?: string | null;
  lastRunResult?: ScraperStatusLastRunResult;
  isRunning: boolean;
  nextRunAt: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetMoviesParams = {
  page?: number;
  /**
   * Content type filter
   */
  type?: GetMoviesType;
  genre?: string;
  search?: string;
};

export type GetMoviesType = (typeof GetMoviesType)[keyof typeof GetMoviesType];

export const GetMoviesType = {
  film: "film",
  anime: "anime",
  all: "all",
} as const;

export type GetMovieDetailParams = {
  /**
   * Resolve direct video URLs (bypasses ad redirects). Default true.
   */
  resolveVideo?: boolean;
};