# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: KlikFilm Scraper API - Film & Anime
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: movies
    description: Movie & Anime scraper operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /movies:
    get:
      operationId: getMovies
      tags: [movies]
      summary: Get movies / anime list
      description: Scrape and return films or anime from klikfilm.web.id
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: type
          in: query
          required: false
          description: Content type filter
          schema:
            type: string
            enum: [film, anime, all]
            default: all
        - name: genre
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of movies / anime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoviesResponse"
        "500":
          description: Scraping error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /movies/{slug}:
    get:
      operationId: getMovieDetail
      tags: [movies]
      summary: Get movie / anime detail
      description: Scrape and return detailed info, including direct video URLs without ads
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: resolveVideo
          in: query
          required: false
          description: Resolve direct video URLs (bypasses ad redirects). Default true.
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Movie / anime detail with direct stream URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Scraping error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /genres:
    get:
      operationId: getGenres
      tags: [movies]
      summary: Get available genres
      description: Returns list of genres scraped from klikfilm.web.id
      responses:
        "200":
          description: List of genres
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenresResponse"
  /scraper/status:
    get:
      operationId: getScraperStatus
      tags: [movies]
      summary: Get daily scraper status
      description: Returns the last run time and next scheduled run for the daily scraper
      responses:
        "200":
          description: Scraper status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScraperStatus"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Movie:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        poster:
          type: string
        year:
          type: string
        rating:
          type: string
        genre:
          type: array
          items:
            type: string
        quality:
          type: string
        type:
          type: string
          enum: [film, anime, all]
        url:
          type: string
      required:
        - title
        - slug
        - poster
        - url
    MovieDetail:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        poster:
          type: string
        year:
          type: string
        rating:
          type: string
        genre:
          type: array
          items:
            type: string
        quality:
          type: string
        type:
          type: string
          enum: [film, anime, all]
        url:
          type: string
        synopsis:
          type: string
        duration:
          type: string
        country:
          type: string
        director:
          type: string
        cast:
          type: array
          items:
            type: string
        streamUrls:
          type: array
          items:
            type: string
          description: Direct video URLs (MP4/M3U8), ad-free
      required:
        - title
        - slug
        - poster
        - url
    MoviesResponse:
      type: object
      properties:
        movies:
          type: array
          items:
            $ref: "#/components/schemas/Movie"
        totalPages:
          type: integer
        currentPage:
          type: integer
      required:
        - movies
        - totalPages
        - currentPage
    GenresResponse:
      type: object
      properties:
        genres:
          type: array
          items:
            type: string
      required:
        - genres
    ScraperStatus:
      type: object
      properties:
        lastRunAt:
          type: string
          nullable: true
        lastRunResult:
          type: object
          nullable: true
          properties:
            films:
              type: integer
            anime:
              type: integer
        isRunning:
          type: boolean
        nextRunAt:
          type: string
      required:
        - isRunning
        - nextRunAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message