# lib/api-zod/src/generated/types/movieDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KlikFilm Scraper API - Film & Anime
 * OpenAPI spec version: 0.2.0
 */
import type { MovieDetailType } from "./movieDetailType";

export interface MovieDetail {
  title: string;
  slug: string;
  poster: string;
  year?: string;
  rating?: string;
  genre?: string[];
  quality?: string;
  type?: MovieDetailType;
  url: string;
  synopsis?: string;
  duration?: string;
  country?: string;
  director?: string;
  cast?: string[];
  /** Direct video URLs (MP4/M3U8), ad-free */
  streamUrls?: string[];
}