# lib/api-zod/src/generated/types/moviesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KlikFilm Scraper API - Film & Anime
 * OpenAPI spec version: 0.2.0
 */
import type { Movie } from "./movie";

export interface MoviesResponse {
  movies: Movie[];
  totalPages: number;
  currentPage: number;
}