# lib/api-zod/src/generated/types/scraperStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KlikFilm Scraper API - Film & Anime
 * OpenAPI spec version: 0.2.0
 */
import type { ScraperStatusLastRunResult } from "./scraperStatusLastRunResult";

export interface ScraperStatus {
  lastRunAt?: string | null;
  lastRunResult?: ScraperStatusLastRunResult;
  isRunning: boolean;
  nextRunAt: string;
}